breed-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.15.1
Pillow==10.4.0
python-multipart==0.0.9
numpy==1.26.4
